+++ libs/spark/src/Collapse.ts
@@ -1,0 +1,6 @@
+import Collapse from '@material-ui/core/Collapse';
+import type { CollapseProps } from '@material-ui/core/Collapse';
+
+export default Collapse;
+
+export type { CollapseProps };