+++ libs/spark/src/ScopedCssBaseline.ts
@@ -1,0 +1,6 @@
+import ScopedCssBaseline from '@material-ui/core/ScopedCssBaseline';
+import type { ScopedCssBaselineProps } from '@material-ui/core/ScopedCssBaseline';
+
+export default ScopedCssBaseline;
+
+export type { ScopedCssBaselineProps };