+++ libs/spark/src/Skeleton.ts
@@ -1,0 +1,6 @@
+import Skeleton from '@material-ui/lab/Skeleton';
+import type { SkeletonProps } from '@material-ui/lab/Skeleton';
+
+export default Skeleton;
+
+export type { SkeletonProps };