+++ libs/spark/src/styled.ts
@@ -1,3 +1,6 @@
 import styled from '@material-ui/core/styles/styled';
+import type { StyledProps } from '@material-ui/core/styles/styled';
 
 export default styled;
+
+export type { StyledProps };