+++ libs/spark/src/withStyles.ts
@@ -1,3 +1,6 @@
 import withStyles from '@material-ui/core/styles/withStyles';
+import type { WithStyles } from '@material-ui/core/styles/withStyles';
 
 export default withStyles;
+
+export type { WithStyles };